+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Repository and Task objects have many additional attributes
+
+### Changed
+- Changed formatting of task error text; now includes a header with the ID of
+  the failed task
+
 ### Fixed
 - Fixed a crash on Python 2.6